Add header_checks dir and changed prepare_static_headers and patch_static_headers to use some unified functions
